visualizer/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pymupdf
pdfminer.six
